download fixes
(cherry picked from commit ecb42eeab0ad0968fcc2df4b04b5cbf3016c161d)

+++ app/src/main/java/com/github/andreyasadchy/xtra/repository/OfflineRepository.kt
@@ -3,12 +3,8 @@ package com.github.andreyasadchy.xtra.repository
 import android.content.Context
 import com.github.andreyasadchy.xtra.db.BookmarksDao
 import com.github.andreyasadchy.xtra.db.LocalFollowsChannelDao
-import com.github.andreyasadchy.xtra.db.RequestsDao
 import com.github.andreyasadchy.xtra.db.VideosDao
 import com.github.andreyasadchy.xtra.model.offline.OfflineVideo
-import com.github.andreyasadchy.xtra.model.offline.Request
-import com.github.andreyasadchy.xtra.ui.download.DownloadService
-import com.github.andreyasadchy.xtra.util.DownloadUtils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
@@ -20,7 +16,6 @@ import javax.inject.Singleton
 @Singleton
 class OfflineRepository @Inject constructor(
         private val videosDao: VideosDao,
-        private val requestsDao: RequestsDao,
         private val localFollowsChannelDao: LocalFollowsChannelDao,
         private val bookmarksDao: BookmarksDao) {
 
@@ -44,12 +39,12 @@ class OfflineRepository @Inject constructor(
 
     suspend fun deleteVideo(context: Context, video: OfflineVideo) = withContext(Dispatchers.IO) {
         video.videoId?.let {
-            if (it.isNotBlank() && bookmarksDao.getByVideoId(it) == null) {
+            if (it.isNotBlank() && videosDao.getByVideoId(it).isEmpty() && bookmarksDao.getByVideoId(it) == null) {
                 File(context.filesDir.path + File.separator + "thumbnails" + File.separator + "${it}.png").delete()
             }
         }
         video.channelId?.let {
-            if (it.isNotBlank() && localFollowsChannelDao.getByUserId(it) == null && bookmarksDao.getByUserId(it).isEmpty()) {
+            if (it.isNotBlank() && getVideosByUserId(it).isEmpty() && bookmarksDao.getByUserId(it).isEmpty() && localFollowsChannelDao.getByUserId(it) == null) {
                 File(context.filesDir.path + File.separator + "profile_pics" + File.separator + "${it}.png").delete()
             }
         }
@@ -68,21 +63,5 @@ class OfflineRepository @Inject constructor(
 
     suspend fun deletePositions() = withContext(Dispatchers.IO) {
         videosDao.deletePositions()
-    }
-
-    suspend fun resumeDownloads(context: Context) = withContext(Dispatchers.IO) {
-        requestsDao.getAll().forEach {
-            if (DownloadService.activeRequests.add(it.offlineVideoId)) {
-                DownloadUtils.download(context, it)
-            }
-        }
-    }
-
-    suspend fun saveRequest(request: Request) = withContext(Dispatchers.IO) {
-        requestsDao.insert(request)
-    }
-
-    suspend fun deleteRequest(request: Request) = withContext(Dispatchers.IO) {
-        requestsDao.delete(request)
     }
 }
